Minor reorg. Split the reducer to its own file.
