+++ .gitignore
@@ -1,3 +1,4 @@
 lib
 node_modules
 coverage
+*.log
