+++ src/index.js
@@ -1,9 +1,9 @@
+export syncHistoryWithStore from './sync'
+export { LOCATION_CHANGE, routerReducer } from './reducer'
+
 export {
   UPDATE_LOCATION,
   push, replace, go, goBack, goForward,
   routeActions
 } from './actions'
-
 export routerMiddleware from './middleware'
-
-export { LOCATION_CHANGE, routerReducer, syncHistoryWithStore } from './sync'
